Passport
Signing up with an email and setting a password, then signing out. Then signing in with DISCORD will continue to sign into that initial account if the emails are the same.

This means we can use any service, and LINK accounts via email for easy sign in, but it would probably still be a good idea to ask them for a password if they SIGN UP using another service.